Fix trouble with uniqu_ptr deleting reference content.
It is not necessary to use a unique_ptr (also always using .release shows
this), so there can be used a raw pointer instead.

Here we fix usage of _aux an also of back pointing _ac. Especially on _ac
it does not make any sense, since this pointer is stored in many locations.

refs #138