DNS server stop independent of webserver type
In some situations (not passing a WebServer instance from external) the
DNS server was not stopped in _stopPortal(). Maybe the intention was here not
to stopped an externally passed WebServer but I see no use-case for the
dns server here.

Fixed this to remove the enum handling in the next step.